feat(replay): add proof of concept
This repository is a place to try out a new idea. Part of that idea
requires constructing `tracing` `Metadata` on the fly. This is somewhat
complicated because certain tracing methods require a reference to
metadata with a `'static` lifetime.

This initial proof of concept shows that it can be done with only a very
small amount of unsafe code. The unsafe code can probably be removed at
the expense of using `Box::leak` to create true static lifetime
references. That does involve leaking memory (as the name implies), but
for what we're trying to do, we'd probably never be able to free that
memory anyway, so it doesn't make much difference.