feat(replay): add dynamic span creation
It is now possible to create a span dynamically. It uses the same
metadata store and metadata creation as for the event.

It seems that simply creating a span is enough for the subscriber to
receive a `new` event for that span.

Once a span is created, the rest of the span lifecycle works as normal.
So the subscriber receives the `enter`, `exit`, and `close` events as
normal.