feat(replay): call dispatcher span methods explicitly
In the previous commit, support for spans was added. However, only for
creating a new span via the `tracing` crate's own `Span::new` method and
then calling methods on the `Span` object that is returned.

This isn't ideal for the `tracing-replay` use-case where we want to
enter and exit spans without holding on to an `Entered` guard for each
span that is currently active.

This has now been replaced to only deal with `span::Id` objects and call
the `new_span`, `enter`, `exit`, and `try_close` methods on the
`Dispatcher` directly.