feat(rec): add new crate `tracing-rec`
This is the initial (incomplete) implementation of the `tracing-rec`
crate. This crate provides a `tracing-subscriber` layer which will
record (hence "rec") traces it receives.

The idea is to record data as close as possible to what the layer
receives and then serialize it into a format that can later be read by
`tracing-replay` and replayed (hence "replay") into a `tracing`
dispatcher.

The initial implementation records calls to all the methods on `Layer`
which are needed to reproduce the same sequence elsewhere. They are
serialized into JSON (because it allows for easy visual debugging) and
written to `stdout`.

Two examples, `events` and `spans` have been included which record
events and spans respectively so that all the methods on `Layer` are
invoked between the two of them.