feat(rec): add new crate `tracing-rec` (#2)
This is the initial (incomplete) implementation of the `tracing-rec`
crate. This crate provides a `tracing-subscriber` layer which will
record (hence "rec") traces it receives.

The idea is to record data as close as possible to what the layer
receives and then serialize it into a format that can later be read by
`tracing-replay` and replayed (hence "replay") into a `tracing`
dispatcher.

The initial implementation records calls to `Layer::register_callsite`
and `Layer::on_event`. The contents are then serialized into JSON
(because it allows easy visual debugging) and written to `stdout`.

An example called `events` is also included which records some events
(and a span to show the parent field) which are then output by the `Rec`
layer.