feat(rec): add support for recording spans (#4)
As well as recording events, we wish to record spans.

Extends the `tracing-rec` crate to also support recording spans. This
completes the implementation of all the methods on `Layer` which are
needed to reproduce the same sequence of traces elsewhere.

A new example, `spans`, has also been added so that all the methods on
`Layer` are invoked between that and the already existing `events`
example.