feat(rec): add recording metadata
As well as the information required to reproduce a traces, certain
metadata is required so that the reproduced traces behave correctly. As
no context information (current active span) is recorded, it is
important that the traces are recreated in threads which also mimic the
threads in the original recording.

This change embeds the `Trace` enum in a `TraceRecord` struct which also
contains `RecordMeta`. The metadata contains a timestamp as well as the
thread ID (as a debug string, which is the only API that `stable`
offers) and the name of the thread.

This will give `tracing-replay` enough information to reproduce the
context for each trace.

A new example called `threads` has also been added which interleaves
spans and events in 2 different threads.