feat(replay): add support to replay events from file (#6)
With the tracing recording file which can be created by `tracing-rec`,
the next step is to replay the contents back through the `tracing`
dispatcher.

This change adds initial support for that, but only for Events. Events
are replayed with up to 9 fields. Due to the nature of the tracing
structs representing fields, it seems to not be possible to support an
unlimited number of fields (which tracing now does), but support should
be added up to the old tracing limit of 32 fields at some stage.