refac(replay): factor out `ValueSet` creation
Creating a `ValueSet` is one of the trickier parts of the
`tracing-replay` crate. In the previous implementation, almost exactly
the same code was present in 3 different places. Once for each of new
span, event, and record.

This change moves that functionality into a new `DispatchProxy` trait
and three implementations for each of the three scenarios. This reduces
duplication and helps make the surrounding code clearer.

With this change, we also extend the value set creation to include up to
32 fields. Unfortunately we are still limited by the number of fields
which can be created, but with the current `tracing` API, this seems to
be inevitable. At least we can now support as many fields as `tracing`
prior to 0.1.40 (when the field count limit was removed).