feat(replay): add support for multi-threaded recordings
Tracing depends on thread-local storage to determine the (contextual)
parent of an event or span. This means that not only is the order of
traces important, but also in which thread they occurred. If a span is
entered in one thread and then an event is recorded in another, that
span should not be considered the parent of the event.

The recordings produced by tracing-rec already include thread
information in the recording metadata, however tracing-replay didn't
make use of this.

This change adds support for replaying traces in the correct thread. As
the recording is processed, a map of threads is kept, if an unknown
thread is found, then a new thread with the same name is spawned with
the receiving end of an MPSC channel. Recorded traces for that thread
are then sent through the the channel to be dispatched from there.

No attempt is made to keep the thread Id the same as in the recording,
but will be a 1-to-1 mapping between recorded thread Id and replayed
thread Id.

Some operations are still performed on the replay thread to avoid
contention. Specifically, tracing Metadata is created and stored and the
map from recording `SpanId` to callsite Id (used to look up tracing
Metadata) is all kept on the main thread.

The recording `SpanId` to tracing `span::Id` mapping, however, is shared
across all threads. This is because the mapping isn't known until the
new span trace is dispatched, which should happen on the right thread.
To accomodate this, the span id store first inserts a `Pending` token as
a value in the store, if another thread needs to access the span Id
mapping and finds it pending, it knows to wait for it to become
available - a different situation from a broken recording file which
doesn't contain the new span trace for some later span lifecycle trace.

This implementation is only feature complete, it requires significant
refactoring.