feat(replay): replay trace records by timestamp
The recording produced by `tracing-rec` includes the timestamp when each
trace was recorded. This change to `tracing-replay` respects the rate at
which the traces are recorded when replaying.

This means that if a trace was recorded 210 milliseconds after the very
first trace was recorded, the replayer will not replace that trace until
the same 210 milliseconds has passed since the first trace record was
replayed.

The delay is performed in the thread that will replay the trace. This
implies that the traces for a specific thread must have been recorded in
order, but that it is acceptable for traces from different threads to be
out of order with respect to one another.

No attempt has been made to replay with exact times.